Add UpdateReceiver project to solution
Added a new project named "UpdateReceiver" to the solution. This project includes a program that triggers actions like sending messages, polls and chat actions for testing purposes via the telegram bot. The program subscribes to different types of updates and outputs them to console. The UpdateReceiver project was added to test and demonstrate the reception and processing of these updates with our RxTelegram Bot.